fix(latency): address review findings on the capture API
Follow-ups from review of the reworked surface, plus the semantic half
of the merge with main:

- `skip` bindings take `usize`: main's #859 moved the core combinator
  off `u32`, and the merge was textually clean but did not compile.
- A second `windows()` stream on one `LatencyHandle` now panics at
  wiring time. The read is destructive, so two streams silently stole
  samples from each other — counts split arbitrarily, no error.
- The octave ceiling returns to 2^34 (17.2 s). 2^32 (4.29 s) does not
  clear the end-to-end row its own rationale names: a WAN reconnect
  stall runs seconds to tens of seconds, and a saturating total reads
  as a plausible number rather than as an overflow. The 7.7 KB
  per-stage cost is affordable now that `StageStats` is not `Copy`.
- Python `stats.total` mirrors the Rust `n < 2` guard instead of
  fabricating a `from == to` row for a single-stage schema.
- Python's fused stamp node is labelled `stamp_all`, matching its Rust
  twin, so diagnostics point at the call the user wrote.
- `hop_dict` reads from the new `StageStats::summary_unlabelled`, so
  the sentinel rule and quantile set exist in one place and the two
  surfaces cannot drift.
- The "every entry point has an `_if` variant" doc bullet is scoped to
  the named forms; `stamp_as`/`stamp_all` toggle with `mode="off"`.
- Precise-stamp tests no longer assert strict monotonicity across
  back-to-back clock reads, which collide on coarse monotonic clocks
  (24-50 MHz ARM counters).

+++ crates/wingfoil-python/src/graph.rs
@@ -428,7 +428,7 @@ impl PyStream {
     }
 
     /// Suppress the first `n` values, then pass every later value through.
-    pub fn skip(&self, n: u32) -> PyStream {
+    pub fn skip(&self, n: usize) -> PyStream {
         self.wrap(self.stream.skip(n))
     }
